move parse_team_model_from_str to bpl_utils
